refactor: extract inboxSupportsReplyTo to shared utility in inboxUtils
